update and refactor add readme
