+++ requirements.txt
@@ -1,3 +1,13 @@
-langchain==0.1.0
 faiss-cpu==1.7.4
-sentence_transformers==2.5.1
+langchain==0.3.4
+langchain-anthropic==0.2.3
+langchain-cohere==0.1.0
+langchain-community==0.3.3
+langchain-core==0.3.13
+langchain-ollama==0.2.0
+langchain-openai==0.2.4
+langchain-text-splitters==0.3.0
+pandas==2.1.1
+PyPDF2==3.0.1
+PyYAML==6.0.1
+rouge-score==0.1.2
