+++ rfpgo/checklist/prompts.py
@@ -61,64 +61,3 @@ Response with outline only, nothing else.
 </questions>
 
 <outline_update>"""
-
-
-
-## below - not showing value
-checklist_from_page_summaries = """\
-The following are summaries of pages in a Request for Proposal document. \
-Review them carefully for information about what is required for drafting a response to this document. \
-Consolidate the information into a detailed outline of the sections required in a response. \
-Include information about response format, important dates, and any other relevant information.
-
-Document:
-{document}
-
-Outline: """
-
-format_sections = """The following outline contains information about the sections in a proposal. \
-Format this information into this standard format:
-
-1. Section name
-- Section detail 1
-- Section detail 2
-
-2. Section name
-...
-
-General requirements:
-- Requirement 1
-- Requirement 2
-...
-
-{document}
-"""
-
-get_sections = """The following document contains information about sections in a proposal. \
-Return a list of the sections, one per line. \
-No additional information is needed. \
-Use the following format:\n\n
-
-1. First section name\n
-2. Second section name\n
-...
-
-{document}"""
-
-
-rag_sections = """What are the sections required for a response to this proposal?"""
-
-rag_sections_detail = """What are the details for the {document} section?"""
-
-get_sections_detail = """The following document contains information about \
-the {section_name} section of a proposal. \
-Return the requirements for drafting this section as a brief summary.
-
-Document:\n
-{document}
-Summary:\n"""
-
-
-section_questions = """Drafting the {document_1} section has the following requirements:\n\
-{document_2}\n\n\
-In drafting the section, what questions would you need to answer?"""
